Move unit tests into submodule
